# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Opora API
  version: 0.3.0
  description: Authenticated API for the Opora student document workspace.
servers:
  - url: http://localhost:8080
...
paths:
...
  /api/v1/meetings:
    get:
      operationId: listMeetings
      tags: [Meetings]
      security: [{ sessionCookie: [] }]
      responses:
        "200": { description: Meetings for visible students, content: { application/json: { schema: { type: object, properties: { items: { type: array, items: { $ref: "#/components/schemas/CouncilMeeting" } } } } } } }
  /api/v1/students/{studentId}/meetings:
    parameters: [{ $ref: "#/components/parameters/StudentID" }]
    get:
      operationId: listStudentMeetings
      tags: [Meetings]
      security: [{ sessionCookie: [] }]
      responses:
        "200": { description: Student meetings, content: { application/json: { schema: { type: object, properties: { items: { type: array, items: { $ref: "#/components/schemas/CouncilMeeting" } } } } } } }
    post:
      operationId: createStudentMeeting
      tags: [Meetings]
      security: [{ sessionCookie: [], csrfHeader: [] }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: "#/components/schemas/CreateCouncilMeeting" } } } }
      responses:
        "201": { description: Meeting created, content: { application/json: { schema: { $ref: "#/components/schemas/CouncilMeeting" } } } }
        "403": { $ref: "#/components/responses/Error" }
  /api/v1/meetings/{meetingId}:
    parameters: [{ name: meetingId, in: path, required: true, schema: { type: string, format: uuid } }]
    get:
      operationId: getMeeting
      tags: [Meetings]
      security: [{ sessionCookie: [] }]
      responses:
        "200": { description: Tenant-scoped meeting, content: { application/json: { schema: { $ref: "#/components/schemas/CouncilMeeting" } } } }
        "403": { $ref: "#/components/responses/Error" }
    patch:
      operationId: updateMeeting
      tags: [Meetings]
      security: [{ sessionCookie: [], csrfHeader: [] }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: "#/components/schemas/UpdateCouncilMeeting" } } } }
      responses:
        "200": { description: Meeting updated or completed, content: { application/json: { schema: { $ref: "#/components/schemas/CouncilMeeting" } } } }
        "403": { $ref: "#/components/responses/Error" }
  /api/v1/tasks:
    get:
      operationId: listMyTasks
      tags: [Tasks]
      security: [{ sessionCookie: [] }]
      responses:
        "200": { description: Tasks assigned to current user, content: { application/json: { schema: { type: object, properties: { items: { type: array, items: { $ref: "#/components/schemas/StudentTask" } } } } } } }
  /api/v1/students/{studentId}/tasks:
    parameters: [{ $ref: "#/components/parameters/StudentID" }]
    get:
      operationId: listStudentTasks
      tags: [Tasks]
      security: [{ sessionCookie: [] }]
      responses:
        "200": { description: Student tasks, content: { application/json: { schema: { type: object, properties: { items: { type: array, items: { $ref: "#/components/schemas/StudentTask" } } } } } } }
    post:
      operationId: createStudentTask
      tags: [Tasks]
      security: [{ sessionCookie: [], csrfHeader: [] }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: "#/components/schemas/CreateStudentTask" } } } }
      responses:
        "201": { description: Task created, content: { application/json: { schema: { $ref: "#/components/schemas/StudentTask" } } } }
        "403": { $ref: "#/components/responses/Error" }
...
components:
  securitySchemes:
    sessionCookie: { type: apiKey, in: cookie, name: opora_session }
    csrfHeader: { type: apiKey, in: header, name: X-CSRF-Token }
  parameters:
    StudentID: { name: studentId, in: path, required: true, schema: { type: string, format: uuid } }
...
  responses:
...
    Error:
      description: Stable public error envelope.
      content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } }
  schemas:
...
    CouncilMeeting:
      type: object
      required: [id, studentId, studentName, scheduledAt, status, subject, participantIds, participantNames, createdAt, updatedAt]
      properties:
        id: { type: string, format: uuid }
        studentId: { type: string, format: uuid }
        studentName: { type: string }
        scheduledAt: { type: string, format: date-time }
        status: { type: string, enum: [planned, completed, cancelled] }
        subject: { type: string }
        questions: { type: [string, "null"] }
        notes: { type: [string, "null"] }
        decision: { type: [string, "null"] }
        recommendations: { type: [string, "null"] }
        participantIds: { type: array, items: { type: string, format: uuid } }
        participantNames: { type: array, items: { type: string } }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    CreateCouncilMeeting:
      type: object
      additionalProperties: false
      required: [scheduledAt, subject]
      properties:
        scheduledAt: { type: string, format: date-time }
        subject: { type: string, minLength: 1, maxLength: 500 }
        questions: { type: string, maxLength: 5000 }
        notes: { type: string, maxLength: 5000 }
        participantIds: { type: array, uniqueItems: true, items: { type: string, format: uuid } }
    UpdateCouncilMeeting:
      allOf:
        - $ref: "#/components/schemas/CreateCouncilMeeting"
        - type: object
          required: [status, decision, recommendations]
          properties:
            status: { type: string, enum: [planned, completed, cancelled] }
            decision: { type: string, maxLength: 5000 }
            recommendations: { type: string, maxLength: 5000 }
    StudentTask:
      type: object
      required: [id, title, assigneeUserId, assigneeName, status, createdAt, updatedAt]
      properties:
        id: { type: string, format: uuid }
        studentId: { type: [string, "null"], format: uuid }
        studentName: { type: [string, "null"] }
        meetingId: { type: [string, "null"], format: uuid }
        title: { type: string }
        description: { type: [string, "null"] }
        assigneeUserId: { type: string, format: uuid }
        assigneeName: { type: string }
        dueAt: { type: [string, "null"], format: date-time }
        status: { type: string, enum: [todo, done] }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    CreateStudentTask:
      type: object
      additionalProperties: false
      required: [title, assigneeUserId]
      properties:
        title: { type: string, minLength: 1, maxLength: 255 }
        description: { type: string, maxLength: 3000 }
        assigneeUserId: { type: string, format: uuid }
        dueAt: { type: string, format: date-time }
        meetingId: { type: [string, "null"], format: uuid }
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string }
            message: { type: string }
            requestId: { type: string }